requirements.txt
Flask
flask_marshmallow
marshmallow_sqlalchemy
marshmallow
Flask-SQLAlchemy
mysqlclient
PyJWT
